Rename files and use hex crate
