+++ solana-programs/staking-bridge/programs/staking-bridge/src/wormhole.rs
@@ -65,14 +65,6 @@ pub fn check_wormhole_token_accounts(
     Ok(())
 }
 
-fn hex_string_to_array(hex_string: &str) -> Vec<u8> {
-    hex_string
-        .as_bytes()
-        .chunks(2)
-        .map(|chunk| u8::from_str_radix(std::str::from_utf8(chunk).unwrap_or("00"), 16).unwrap_or(0))
-        .collect()
-}
-
 pub fn check_wormhole_pdas(
   accounts: &mut PostWormholeMessage,
   config_bump: u8,
@@ -104,7 +96,7 @@ pub fn check_wormhole_pdas(
     }
 
     let (wrapped_mint_pda, wrapped_mint_pda_bump) = Pubkey::find_program_address(
-        &[b"wrapped".as_ref(), &ETH_CHAIN_ID.to_be_bytes()[..2], hex_string_to_array(ETH_AUDIO_TOKEN_ADDRESS_PADDED_32_BYTES).as_ref()],
+        &[b"wrapped".as_ref(), &ETH_CHAIN_ID.to_be_bytes()[..2], hex::decode(ETH_AUDIO_TOKEN_ADDRESS_PADDED_32_BYTES).unwrap().as_ref()],
         program_id.key
     );
     if *wrapped_mint.key != wrapped_mint_pda || wrapped_mint_bump != wrapped_mint_pda_bump {
@@ -226,7 +218,7 @@ pub fn execute_wormhole_transfer(
     let system_program = &accounts.system_program;
 
     let mut target_address = [0u8; 32];
-    target_address[..32].copy_from_slice(&hex_string_to_array(ETH_RECIPIENT_ADDRESS_PADDED_32_BYTES));
+    target_address[..32].copy_from_slice(&hex::decode(ETH_RECIPIENT_ADDRESS_PADDED_32_BYTES).unwrap());
     let target_chain = ETH_CHAIN_ID;
 
     // https://github.com/wormhole-foundation/wormhole/blob/main/solana/modules/token_bridge/program/src/lib.rs#L107
